+++ pyxelrest_auto_update.py
@@ -6,13 +6,25 @@ import subprocess
 import os
 import time
 import sys
+import yaml
 import logging
+import logging.config
 import logging.handlers
 
-default_log_file_path = os.path.join(os.getenv('APPDATA'), 'pyxelrest', 'logs', 'pyxelrest_auto_update.log')
-logging.basicConfig(format='%(asctime)s [%(levelname)s] %(message)s',
-                    handlers=[logging.handlers.TimedRotatingFileHandler(default_log_file_path, when='D')],
-                    level=logging.DEBUG)
+
+logging_configuration_file_path = os.path.join(os.getenv('APPDATA'), 'pyxelrest', 'configuration',
+                                               'auto_update_logging.ini')
+if os.path.isfile(logging_configuration_file_path):
+    with open(logging_configuration_file_path, 'r') as config_file:
+        log_config_dict = yaml.load(config_file)
+        logging.config.dictConfig(log_config_dict)
+else:
+    default_log_file_path = os.path.join(os.getenv('APPDATA'), 'pyxelrest', 'logs', 'pyxelrest_auto_update.log')
+    logging.basicConfig(format='%(asctime)s [%(levelname)s] %(message)s',
+                        handlers=[logging.handlers.TimedRotatingFileHandler(default_log_file_path, when='D')],
+                        level=logging.DEBUG)
+    logging.warning('Logging configuration file ({0}) cannot be found. Using default logging configuration.'.format(
+        logging_configuration_file_path))
 
 
 class PyxelRestUpdater: